# [Merged by Bors] - chore: remove unnecessary term or tactic `set_option in` with backwards-compatibility flags