--- a/Mathlib/Analysis/Convex/Uniform.lean
+++ b/Mathlib/Analysis/Convex/Uniform.lean
@@ -104,7 +104,7 @@ theorem exists_forall_closed_ball_dist_add_le_two_sub (hε : 0 < ε) :
       refine sub_le_sub_left ?_ _
       ring_nf
       rw [← mul_div_cancel₀ δ three_ne_zero]
-      set_option tactic.skipAssignedInstances false in norm_num
+      norm_num
       -- Porting note: these three extra lines needed to make `exact` work
       have : 3 * (δ / 3) * (1 / 3) = δ / 3 := by linarith
       rw [this, mul_comm]