--- a/Mathlib/LinearAlgebra/CrossProduct.lean
+++ b/Mathlib/LinearAlgebra/CrossProduct.lean
@@ -102,7 +102,7 @@ theorem cross_self (v : Fin 3 → R) : v ×₃ v = 0 := by
 @[simp 1100, nolint simpNF] -- Porting note: increase priority so that the LHS doesn't simplify
 theorem dot_self_cross (v w : Fin 3 → R) : v ⬝ᵥ v ×₃ w = 0 := by
   rw [cross_apply, vec3_dotProduct]
-  set_option tactic.skipAssignedInstances false in norm_num
+  norm_num
   ring
 #align dot_self_cross dot_self_cross