@grunweg 2024-06-11
Lesson learned: these options can be performance-sensitive. Reverted the `ScaleRoots` change and added a comment.
@MichaelStollBayreuth Can I ask you to run your script on the benchmarking run above? I'd like to know if there are any other files with significant regressions. Thank you (that script is really nice - I wish the speed-center also allowed sorting by the absolute change)!