@MichaelStollBayreuth 2024-06-11
```text
General information:
  build:                                               +19.345 * 10⁹ (+0.0159%)

5 files got slower by at least 10⁹ instructions:
  Mathlib.RingTheory.Polynomial.ScaleRoots:            +11.369 * 10⁹ (+49.6%)
  Mathlib.Analysis.SpecialFunctions.Trigonometric.Bounds:
                                                       +4.1656 * 10⁹ (+6.95%)
  Mathlib.Topology.Homotopy.Path:                      +2.7929 * 10⁹ (+3.84%)
  Mathlib.Data.Finset.Pointwise:                       +1.8385 * 10⁹ (+1.25%)
  Mathlib.RingTheory.Localization.Basic:               +1.4865 * 10⁹ (+1.33%)

1 file got slower by at least 10%:
  Mathlib.RingTheory.Polynomial.ScaleRoots:                           +49.6%

2 files got faster by at least 10⁹ instructions:
  Mathlib.Analysis.NormedSpace.PiTensorProduct.InjectiveSeminorm:
                                                       -3.9465 * 10⁹ (-1.32%)
  Mathlib.Algebra.Category.ModuleCat.ChangeOfRings:    -1.1021 * 10⁹ (-0.278%)

No file got faster by at least 10%.
```