@grunweg 2024-06-11
I looked at the files which slowed down by more than 10^9 instructions. `ScaleRoots` clearly needs to be reverted (with a comment). The simp call in `Data.Finset.Pointwise` speeds up a lot with that option (250 ->50ms locally), hence I also reverted that. In `Trigonometric/Bounds`, the situation seems murkier (the refine's are slow either way, but perhaps go from 850->800ms)... but I've reverted to try another benchmarking run.
The other changes seems inconclusive or a wash; I tend to leave those in.

So, let's re-benchmark this!