@MichaelStollBayreuth 2024-06-11
<b>General information:</b><tt>
 build:                                               +1.0814 * 10⁹ (+0.000889%)
</tt>

<b>1 file got slower by at least 10⁹ instructions:</b>
<tt> Mathlib.Topology.Homotopy.Path:                      +2.7944 * 10⁹ (+3.84%)
</tt>

<b>No file got slower by at least 10%.</b>

<b>3 files got faster by at least 10⁹ instructions:</b>
<tt> Mathlib.Analysis.NormedSpace.PiTensorProduct.InjectiveSeminorm:
                                                       -2.1722 * 10⁹ (-0.726%)
  Mathlib.Analysis.SpecialFunctions.Trigonometric.Bounds:
                                                       -1.9068 * 10⁹ (-3.18%)
  Mathlib.Algebra.Category.ModuleCat.ChangeOfRings:    -1.1096 * 10⁹ (-0.280%)
</tt>

<b>No file got faster by at least 10%.</b>